backend/requirements.txt
django
djangorestframework
django-cors-headers
pandas
pyyaml
uritemplate
inflection
